requirements.txt
absl-py==1.4.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.7
cloudpickle==2.2.1
cmake==3.27.2
coloredlogs==15.0.1
contourpy==1.1.0
coremltools==6.3.0
cycler==0.11.0
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2023.6.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
grpcio==1.54.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lit==16.0.6
littleutils==0.2.2
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.14.0
onnx-coreml==1.3
onnxruntime==1.15.1
onnxruntime-gpu==1.15.1
opt-einsum==3.3.0
outdated==0.2.2
packaging==23.1
pandas==2.0.2
pandas-flavor==0.6.0
patsy==0.5.3
Pillow==10.0.0
pingouin==0.5.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
statsmodels==0.14.0
submitit==1.4.5
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
timm==0.9.5
torch==2.0.1
torch-tb-profiler==0.4.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
typing==3.7.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
xarray==2023.8.0
